cat_breed_detector_v1/server后端/requirements.txt
flask==2.0.1
werkzeug==2.0.1
pillow>=9.0.0
numpy==1.26.4
opencv-python
# tensorflow 建议用 conda 安装，或者用 pip 安装较新版本